fix(serialization) card fixup for templates
When cards are loaded, the ones with template-tags that refer to other
cards being created on the same `load` can't find their
ids (we first populate them, and then `maybe-upsert-many!` inserts
them at once).

This commit adds a second pass on the inserted cards, narrowing to the
ones that just have been inserted, and resolving the ids again.